+++ packages/web3-eth/test/integration/subscribe.test.ts
@@ -17,30 +17,35 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 import WebSocketProvider from 'web3-providers-ws';
 import { Web3BaseProvider } from 'web3-types';
 /* eslint-disable import/no-named-as-default */
-import Web3Eth from '../../src/index';
-import {
+// eslint-disable-next-line import/no-extraneous-dependencies
+import IpcProvider from 'web3-providers-ipc';
+import Web3Eth, {
 	NewHeadsSubscription,
 	SyncingSubscription,
 	NewPendingTransactionsSubscription,
 	LogsSubscription,
-} from '../../src';
+} from '../../src/index';
 import {
 	getSystemTestProvider,
 	describeIf,
+	isSocket,
 	isWs,
 	createTempAccount,
+	closeOpenConnection,
 } from '../fixtures/system_test_utils';
 
-describeIf(isWs)('subscribe', () => {
+describeIf(isSocket)('subscribe', () => {
 	let web3Eth: Web3Eth;
-	let provider: WebSocketProvider;
+	let provider: WebSocketProvider | IpcProvider;
 
-	beforeAll(async () => {
-		provider = new WebSocketProvider(getSystemTestProvider());
+	beforeAll(() => {
+		provider = isWs
+			? new WebSocketProvider(getSystemTestProvider())
+			: new IpcProvider(getSystemTestProvider());
 	});
 
-	afterAll(() => {
-		provider.disconnect();
+	afterAll(async () => {
+		await closeOpenConnection(web3Eth);
 	});
 
 	afterEach(async () => {