+++ packages/web3-providers-ipc/CHANGELOG.md
@@ -58,3 +58,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   `tsc` compiled files moved to `lib/` directory from `dist/` (#5739)
+-   Refactor to use common SocketProvider class (#5683)