+++ packages/web3-utils/CHANGELOG.md
@@ -64,3 +64,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   `tsc` compiled files moved to `lib/` directory from `dist/` (#5739)
+-   Add SocketProvider class and Eip1193Provider abstract class (#5683)